requirements.txt
playwright>=1.42.0
anthropic>=0.39.0
aiohttp>=3.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
rich>=13.7.0
tenacity>=8.2.0
pytest>=8.0.0
